Login and Signup Page
make a login page added status codes added promises and async awaits in the queries.

need to make data folder from ts to cjs. because right now it is using ts is compiling to mjs and the data folder is needed in the server which is in cjs for importing purposes